# openapi/v1.yaml
openapi: 3.1.0
info:
  title: registerwatch
  description: Gambling regulators' public registers, one schema per register, with history.
  contact:
    name: registerwatch
    url: https://github.com/mishmish004/RegisterWatch
  license:
    name: All rights reserved
    url: https://choosealicense.com/no-permission/
  version: 0.2.0
servers:
- url: /
  description: The deployment serving this document
paths:
...
  /v1/search:
    get:
      tags:
      - search
      summary: Find text in every register's current rows
      operationId: search
      security:
      - ReadToken: []
      - IngestToken: []
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 2
          maxLength: 200
          description: Case-insensitive substring
          title: Q
        description: Case-insensitive substring
      - name: jurisdiction
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
          - type: 'null'
          description: Only these jurisdictions; repeat the parameter for several (`?jurisdiction=gb&jurisdiction=de`)
          title: Jurisdiction
        description: Only these jurisdictions; repeat the parameter for several (`?jurisdiction=gb&jurisdiction=de`)
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Rows returned per table
          default: 20
          title: Limit
        description: Rows returned per table
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchHitPage'
          headers:
            Cache-Control: *id001
            ETag: *id002
            RateLimit-Policy: *id003
            RateLimit: *id004
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '504':
          $ref: '#/components/responses/QueryTimeout'
...
  /search:
    get:
      tags:
      - read
      summary: Search
      operationId: search_search_get
      security:
      - ReadToken: []
      - IngestToken: []
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 2
          title: Q
      - name: jurisdiction
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Jurisdiction
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Search Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Pagination:
      properties:
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
          description: Pass as `cursor` for the next page; null on the last page
        has_more:
          type: boolean
          title: Has More
        limit:
          type: integer
          title: Limit
          description: Page size this response was read with
        total:
          anyOf:
          - type: integer
          - type: 'null'
          title: Total
          description: Matching items; only with `include_total=true` on endpoints that offer it, else
            null
      type: object
      required:
      - next_cursor
      - has_more
      - limit
      title: Pagination
      examples:
      - next_cursor: eyJwIjp7Im8iOjEwMH0sInMiOiI0ZjFkIn0
        has_more: true
        limit: 100
        total: null
...
    Row:
      properties:
        id:
          type: integer
          title: Id
          description: Stable for as long as the row is unchanged; a changed row is a new row (and the
            old one is removed). Pages are keyed on it
        first_seen_snapshot_id:
          type: integer
          title: First Seen Snapshot Id
          description: The complete snapshot that first contained this row
        first_seen_at:
          type: string
          format: date-time
          title: First Seen At
          description: When the first complete snapshot containing this row was recorded
        last_seen_at:
          type: string
          format: date-time
          title: Last Seen At
          description: When the newest complete snapshot still containing it was recorded
        values:
          additionalProperties:
            anyOf:
            - type: string
            - type: integer
            - type: number
            - type: boolean
            - type: string
              format: date-time
            - type: string
              format: date
            - items:
                type: string
              type: array
            - type: 'null'
          type: object
          title: Values
          description: The register's own columns, as published; keys are the table's declared columns
      type: object
      required:
      - id
      - first_seen_snapshot_id
      - first_seen_at
      - last_seen_at
      - values
      title: Row
      examples:
      - &id013
        id: 1187
        first_seen_snapshot_id: 412
        first_seen_at: '2026-10-05T06:01:12Z'
        last_seen_at: '2026-10-07T06:00:41Z'
        values: &id012
          account_number: '102'
          licence_number: 000102-N-000000-000
          status: Pending
          type: Non-Remote
          activity: Bingo
          start_date: null
          end_date: null
...
    SearchHit:
      properties:
        jurisdiction:
          type: string
          title: Jurisdiction
        register:
          type: string
          title: Register
          description: Register slug, also its Postgres schema
        regulator:
          type: string
          title: Regulator
        kind:
          type: string
          enum:
          - licensees
          - blocklist
          title: Kind
        table:
          type: string
          title: Table
        total:
          type: integer
          title: Total
          description: Matching rows in this table
        rows:
          items:
            $ref: '#/components/schemas/Row'
          type: array
          title: Rows
          description: The first `limit` of them
        rows_url:
          type: string
          title: Rows Url
          description: Every match in this table, paged, with the same `total`
      type: object
      required:
      - jurisdiction
      - register
      - regulator
      - kind
      - table
      - total
      - rows
      - rows_url
      title: SearchHit
      examples:
      - &id014
        jurisdiction: CH
        register: ch_esbk
        regulator: Eidgenössische Spielbankenkommission (ESBK)
        kind: blocklist
        table: blocked_domains
        total: 1
        rows:
        - id: 77
          first_seen_snapshot_id: 410
          first_seen_at: '2026-10-05T06:01:12Z'
          last_seen_at: '2026-10-07T06:00:41Z'
          values:
            domain: 0101b00merang-bet.com
            listed_on: '2026-08-25'
        rows_url: /v1/registers/ch_esbk/tables/blocked_domains/rows?q=b00merang&include_total=true
    SearchHitPage:
      properties:
        data:
          items:
            $ref: '#/components/schemas/SearchHit'
          type: array
          title: Data
          description: One entry per table with a match, most matches first
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
      required:
      - data
      - pagination
      title: SearchHitPage
      examples:
      - data:
        - *id014
        pagination:
          next_cursor: null
          has_more: false
          limit: 1
          total: 1
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    FieldError:
      examples:
      - field: limit
        location: query
        message: Input should be less than or equal to 1000
      properties:
        field:
          title: Field
          type: string
        location:
          description: '`query`, `path` or `header`'
          title: Location
          type: string
        message:
          title: Message
          type: string
      required:
      - field
      - location
      - message
      title: FieldError
      type: object
    Problem:
      additionalProperties: true
      description: RFC 9457 Problem Details. Extension members may be added per type.
      examples:
      - detail: 'limit: Input should be less than or equal to 1000'
        errors:
        - field: limit
          location: query
          message: Input should be less than or equal to 1000
        instance: /v1/registers/gb_ukgc/tables/licences/rows
        request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
        status: 400
        title: Invalid parameter
        type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#invalid-parameter
      properties:
        type:
          description: Stable identifier of the problem type; its docs say what to do
          format: uri
          title: Type
          type: string
        title:
          title: Title
          type: string
        status:
          title: Status
          type: integer
        detail:
          description: What went wrong with this request, for a human
          title: Detail
          type: string
        instance:
          description: The request path
          format: uri-reference
          title: Instance
          type: string
        request_id:
          description: Also in the `X-Request-Id` response header and the server's logs
          title: Request Id
          type: string
        errors:
          anyOf:
          - items:
              $ref: '#/components/schemas/FieldError'
            type: array
          - type: 'null'
          default: null
          description: Each invalid parameter, for 400s
          title: Errors
      required:
      - type
      - title
      - status
      - detail
      - instance
      - request_id
      title: Problem
      type: object
  securitySchemes:
    ReadToken:
      type: http
      description: '`READ_TOKEN`. Needed on read routes only when the server sets one; when it does not,
        reads are open. The ingest token reads too.'
      scheme: bearer
    IngestToken:
      type: http
      description: '`INGEST_TOKEN`. Starts and lists ingest runs, and reads. A read token here is a 403.'
      scheme: bearer
  responses:
    BadRequest:
      description: 'A parameter is invalid; `errors` names each one. Types: `invalid-parameter`, `unknown-filter-column`,
        `invalid-cursor`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#invalid-parameter
            title: Invalid parameter
            status: 400
            detail: 'limit: Input should be less than or equal to 1000'
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
            errors:
            - field: limit
              location: query
              message: Input should be less than or equal to 1000
      headers:
        RateLimit-Policy: *id003
        RateLimit: *id004
    Unauthenticated:
      description: 'The bearer token is missing, or is not one this route takes. Types: `unauthenticated`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#unauthenticated
            title: Authentication required
            status: 401
            detail: the bearer token is not valid here
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
      headers:
        WWW-Authenticate:
          description: '`Bearer realm="registerwatch"`, with `error="invalid_token"` when a token was
            sent but is not valid (RFC 6750)'
          schema:
            type: string
        RateLimit-Policy: *id003
        RateLimit: *id004
...
    TooManyRequests:
      description: 'Rate limit exceeded; wait `Retry-After` seconds. Types: `rate-limited`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#rate-limited
            title: Too many requests
            status: 429
            detail: 60 requests per minute for /v1/search; retry in 12 s
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
      headers:
        Retry-After: *id016
        RateLimit-Policy: *id003
        RateLimit: *id004
    InternalError:
      description: 'An unexpected error; quote `request_id` when reporting it. Types: `internal`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#internal
            title: Internal error
            status: 500
            detail: something went wrong on our side; quote request_id if you report it
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
      headers:
        RateLimit-Policy: *id003
        RateLimit: *id004
    ServiceUnavailable:
      description: 'The database cannot be reached (retry after `Retry-After` seconds), or ingest is not
        configured on this deployment. Types: `database-unavailable`, `ingest-disabled`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#database-unavailable
            title: Database unavailable
            status: 503
            detail: the database cannot be reached right now; retry shortly
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
      headers:
        Retry-After: *id016
        RateLimit-Policy: *id003
        RateLimit: *id004
    QueryTimeout:
      description: 'A query ran past the statement timeout and was cancelled; narrow the request, or retry
        after `Retry-After` seconds. Types: `query-timeout`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#query-timeout
            title: Query took too long
            status: 504
            detail: the query was cancelled for taking too long; narrow it and retry
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
      headers:
        Retry-After: *id016
        RateLimit-Policy: *id003
        RateLimit: *id004
    NotModified:
      description: Unchanged since the `ETag` sent in `If-None-Match`; no body
      headers:
        ETag:
          $ref: '#/components/headers/ETag'
        Cache-Control:
          $ref: '#/components/headers/Cache-Control'
        RateLimit-Policy: *id003
        RateLimit: *id004
  headers:
    ETag:
      description: This representation's version (weak). Send it back in `If-None-Match` and the answer
        is a 304 with no body while the data is unchanged. New data shows within 30 seconds of its snapshot.
      schema:
        type: string
      example: W/"9f2c5d0e..."
    Cache-Control:
      description: '`public, max-age=300` when reads are open, `private, max-age=300` when the deployment
        requires a read token; `no-store` on ingest operations'
      schema:
        type: string
...